Implement correct handling of missing labels
