+++ tests/integration/test_headers_and_subtotals.py
@@ -15,6 +15,7 @@ from .fixtures import FRUIT_X_PETS_HS_TOP_BOTTOM
 from .fixtures import CAT_X_DATE_HS_PRUNE
 from .fixtures import CAT_X_NUM_HS_PRUNE
 from .fixtures import PETS_X_FRUIT_HS
+from .fixtures import MISSING_CAT_HS
 
 from cr.cube.crunch_cube import CrunchCube
 
@@ -607,3 +608,19 @@ class TestHeadersAndSubtotals(TestCase):
             axis=1, include_transforms_for_dims=[0, 1]
         ).shape
         np.testing.assert_array_equal(actual, expected)
+
+    def test_missing_cat_hs_labels(self):
+        cube = CrunchCube(MISSING_CAT_HS)
+
+        # Don't expect the missing category "Non voters"
+        expected = [[
+            'Whites',
+            'White college women voters',
+            'White non-college women voters',
+            'White college men voters',
+            'White non-college men voters',
+            'Black voters',
+            'Latino and other voters',
+        ]]
+        actual = cube.labels(include_transforms_for_dims=[0])
+        assert actual == expected
